framework: put CRUDResourceOps in separate file (#154)
